OSIS-6881 : refactor tests
