+++ attestation/tests/views/test_home.py
@@ -39,25 +39,18 @@ from osis_reference_sdk.model.paginated_academic_calendars import PaginatedAcade
 from rest_framework import status
 
 import attestation.views.home
-from base.forms.base_forms import RegistrationIdForm
+from attestation.views.home import _check_display_warning_invoice_not_available_yet
 from base.tests.factories.person import PersonFactory
 from base.tests.factories.student import StudentFactory
-from attestation.views.home import _check_display_warning_invoice_not_available_yet
 
 STUDENT_REGISTRATION_ID = "70531800"
-STUDENT_GLOBAL_ID = "78961314"
 
-MULTIPLE_STUDENT_ERROR = _("A problem was detected with your registration : 2 registration id's are "
-                           "linked to your user.</br> Please contact <a href="
-                           "\"{registration_department_url}\" target=\"_blank\">the Registration "
-                           "department</a>. Thank you.") \
-    .format(registration_department_url=settings.REGISTRATION_ADMINISTRATION_URL)
-
-
-def open_sample_pdf():
-    pdf_path = 'attestation/tests/resources/sample.pdf'
-    with open(pdf_path) as pdf_file:
-        return pdf_file
+MULTIPLE_STUDENT_ERROR = _(
+    "A problem was detected with your registration : 2 registration id's are "
+    "linked to your user.</br> Please contact <a href="
+    "\"{registration_department_url}\" target=\"_blank\">the Registration "
+    "department</a>. Thank you."
+).format(registration_department_url=settings.REGISTRATION_ADMINISTRATION_URL)
 
 
 class HomeTest(TestCase):
@@ -96,7 +89,7 @@ class HomeTest(TestCase):
     def test_without_being_logged(self):
         self.client.logout()
         response = self.client.get(self.url)
-        self.assertRedirects(response, '/login/?next={}'.format(self.url))
+        self.assertRedirects(response, f'/login/?next={self.url}')
 
     def test_with_user_not_a_student(self):
         response = self.client.get(self.url, follow=True)
@@ -176,255 +169,6 @@ class HomeTest(TestCase):
         attestations = attestation_statuses_all_years_json_dict.get('attestationStatusesAllYears')
         self.assertFalse(_check_display_warning_invoice_not_available_yet(2020, attestations))
         self.assertTrue(_check_display_warning_invoice_not_available_yet(2021, attestations))
-
-
-class DownloadAttestationTest(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        year = datetime.date.today().year
-        cls.attestation_type = "test"
-        cls.url = reverse('download_attestation', args=[str(year), cls.attestation_type])
-        cls.person = PersonFactory()
-
-        students_group = Group.objects.create(name='students')
-        cls.permission = Permission.objects.get(codename="is_student")
-        students_group.permissions.add(cls.permission)
-
-    def setUp(self):
-        self.client.force_login(self.person.user)
-
-        self.discriminate_user_patcher = mock.patch(
-            "base.business.student.find_by_user_and_discriminate",
-        )
-        self.mocked_discriminate_user = self.discriminate_user_patcher.start()
-        self.addCleanup(self.discriminate_user_patcher.stop)
-
-    def test_without_being_logged(self):
-        self.client.logout()
-        response = self.client.get(self.url)
-        self.assertRedirects(response, '/login/?next={}'.format(self.url))
-
-    def test_with_user_not_a_student(self):
-        response = self.client.get(self.url, follow=True)
-        self.mocked_discriminate_user.return_value = None
-        self.assertTemplateUsed(response, "access_denied.html")
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_with_multiple_students_assigned_same_person(self):
-        StudentFactory(person=self.person)
-        self.mocked_discriminate_user.side_effect = MultipleObjectsReturned
-        response = self.client.get(self.url, follow=True)
-
-        self.assertTemplateUsed(response, "dashboard.html")
-
-        messages = list(response.context['messages'])
-
-        self.assertEqual(len(messages), 1)
-        self.assertEqual(messages[0].tags, 'error')
-        self.assertEqual(messages[0].message, MULTIPLE_STUDENT_ERROR)
-
-    @patch('attestation.queues.student_attestation.fetch_student_attestation',
-           side_effect=lambda global_id, year, attestation_type, username: None)
-    def test_when_no_attestation_pdf(self, mock_fetch_student_attestation):
-        self.mocked_discriminate_user.return_value = StudentFactory(person=self.person)
-        response = self.client.get(self.url, follow=True)
-
-        self.assertTrue(mock_fetch_student_attestation.called)
-        self.assertTemplateUsed(response, "attestation_home_student.html")
-
-        messages = list(response.context['messages'])
-
-        self.assertEqual(len(messages), 1)
-        self.assertEqual(messages[0].tags, 'error')
-        self.assertEqual(messages[0].message, _('Student attestations'))
-
-    @patch('attestation.queues.student_attestation.fetch_student_attestation',
-           side_effect=lambda global_id, year, attestation_type, username: open_sample_pdf())
-    def test_when_attestation_pdf_fetched(self, mock_fetch_student_attestation):
-        StudentFactory(person=self.person)
-        response = self.client.get(self.url, follow=True)
-
-        self.assertTrue(mock_fetch_student_attestation.called)
-        self.assertEqual(response['Content-Type'], 'application/pdf')
-        self.assertEqual(response['Content-Disposition'], 'attachment; filename="{}.pdf"'.format(self.attestation_type))
-        self.assertEqual(response.content.decode(), str(open_sample_pdf()))
-
-
-class AttestationAdministrationTest(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.url = reverse('attestation_administration')
-        cls.person = PersonFactory()
-
-        cls.permission = Permission.objects.get(codename="is_faculty_administrator")
-        cls.person.user.user_permissions.add(cls.permission)
-
-    def setUp(self):
-        self.client.force_login(self.person.user)
-
-    def test_without_being_logged(self):
-        self.client.logout()
-        response = self.client.get(self.url)
-        self.assertRedirects(response, '/login/?next={}'.format(self.url))
-
-    def test_when_user_is_not_a_faculty_administrator(self):
-        a_person = PersonFactory()
-        self.client.force_login(a_person.user)
-        response = self.client.get(self.url, follow=True)
-
-        self.assertTemplateUsed(response, "access_denied.html")
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_when_faculty_administrator(self):
-        response = self.client.get(self.url, follow=True)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertTemplateUsed(response, "admin/attestation_administration.html")
-
-
-class SelectStudentAttestationTest(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.url = reverse('attestation_admin_select_student')
-        cls.person = PersonFactory()
-
-        cls.permission = Permission.objects.get(codename="is_faculty_administrator")
-        cls.person.user.user_permissions.add(cls.permission)
-
-        Group.objects.create(name='students')
-
-    def setUp(self):
-        self.client.force_login(self.person.user)
-
-        self.academic_calendar_row = AcademicCalendar(**{
-            'reference': 'REFERENCE',
-            'title': 'TITLE',
-            'data_year': 2021,
-            'start_date': datetime.date.today(),
-            'end_date': datetime.date.today()
-        })
-        self.academic_calendar_list_patcher = mock.patch(
-            "reference.services.academic_calendar.AcademicCalendarService.get_academic_calendar_list",
-            return_value=PaginatedAcademicCalendars(**{'results': [self.academic_calendar_row]})
-        )
-        self.mocked_academic_calendar_list = self.academic_calendar_list_patcher.start()
-        self.addCleanup(self.academic_calendar_list_patcher.stop)
-
-    def test_without_being_logged(self):
-        self.client.logout()
-        response = self.client.get(self.url)
-        self.assertRedirects(response, '/login/?next={}'.format(self.url))
-
-    def test_when_user_is_not_a_faculty_administrator(self):
-        a_person = PersonFactory()
-        self.client.force_login(a_person.user)
-        response = self.client.get(self.url, follow=True)
-
-        self.assertTemplateUsed(response, "access_denied.html")
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_get_request(self):
-        response = self.client.get(self.url, follow=True)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertTemplateUsed(response, "admin/attestation_administration.html")
-
-        self.assertIsInstance(response.context['form'], RegistrationIdForm)
-
-    def test_invalid_post_request(self):
-        response = self.client.post(self.url, data={'registration_id': STUDENT_REGISTRATION_ID}, follow=True)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertTemplateUsed(response, "admin/attestation_administration.html")
-        # Message valided in base test
-        self.assertEqual(len(response.context['form'].errors), 1)
-
-    @patch('attestation.queues.student_attestation_status.fetch_json_attestation_statuses', side_effect=lambda x: None)
-    def test_valid_post_request_but_no_attestation(self, mock_fetch_json_attestation_statuses):
-        a_student = StudentFactory(registration_id=STUDENT_REGISTRATION_ID)
-        response = self.client.post(self.url, data={'registration_id': STUDENT_REGISTRATION_ID}, follow=True)
-
-        self.assertTrue(mock_fetch_json_attestation_statuses.called)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertTemplateUsed(response, "attestation_home_admin.html")
-
-        self.assertEqual(response.context['student'], a_student)
-
-        self.assertFalse(response.context['attestations'])
-        self.assertFalse(response.context['current_year'])
-
-    @patch('attestation.queues.student_attestation_status.fetch_json_attestation_statuses',
-           side_effect=lambda x: {'current_year': 2015, 'attestations': [],
-                                  'registration_id': STUDENT_REGISTRATION_ID})
-    def test_valid_post_request(self, mock_fetch_json_attestation_statuses):
-        a_student = StudentFactory(registration_id=STUDENT_REGISTRATION_ID)
-        response = self.client.post(self.url, data={'registration_id': STUDENT_REGISTRATION_ID}, follow=True)
-
-        self.assertTrue(mock_fetch_json_attestation_statuses.called)
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertTemplateUsed(response, "attestation_home_admin.html")
-
-        self.assertEqual(response.context['student'], a_student)
-        self.assertEqual(response.context['current_year'], 2015)
-
-        self.assertFalse(response.context['attestations'])
-
-
-class DownloadStudentAttestation(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        year = datetime.date.today().year
-        cls.attestation_type = "test"
-        cls.url = reverse('attestation_admin_download', args=[STUDENT_GLOBAL_ID, year, cls.attestation_type])
-        cls.person = PersonFactory()
-
-        cls.permission = Permission.objects.get(codename="is_faculty_administrator")
-        cls.person.user.user_permissions.add(cls.permission)
-
-        Group.objects.create(name='students')
-
-    def setUp(self):
-        self.client.force_login(self.person.user)
-
-    def test_without_being_logged(self):
-        self.client.logout()
-        response = self.client.get(self.url)
-        self.assertRedirects(response, '/login/?next={}'.format(self.url))
-
-    def test_when_user_is_not_a_faculty_administrator(self):
-        a_person = PersonFactory()
-        self.client.force_login(a_person.user)
-        response = self.client.get(self.url, follow=True)
-
-        self.assertTemplateUsed(response, "access_denied.html")
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    @patch('attestation.queues.student_attestation.fetch_student_attestation',
-           side_effect=lambda global_id, year, attestation_type, username: None)
-    def test_when_no_attestation_pdf(self, mock_fetch_student_attestation):
-        StudentFactory(person=PersonFactory(global_id=STUDENT_GLOBAL_ID))
-        response = self.client.get(self.url, follow=True)
-
-        self.assertTrue(mock_fetch_student_attestation.called)
-        self.assertTemplateUsed(response, "attestation_home_admin.html")
-
-        messages = list(response.context['messages'])
-
-        self.assertEqual(len(messages), 1)
-        self.assertEqual(messages[0].tags, 'error')
-        self.assertEqual(messages[0].message, _('Student attestations'))
-
-    @patch('attestation.queues.student_attestation.fetch_student_attestation',
-           side_effect=lambda global_id, year, attestation_type, username: open_sample_pdf())
-    def test_when_attestation_pdf_fetched(self, mock_fetch_student_attestation):
-        response = self.client.get(self.url, follow=True)
-
-        self.assertTrue(mock_fetch_student_attestation.called)
-        self.assertEqual(response['Content-Type'], 'application/pdf')
-        self.assertEqual(response['Content-Disposition'], 'attachment; filename="{}.pdf"'.format(self.attestation_type))
-        self.assertEqual(response.content.decode(), str(open_sample_pdf()))
 
 
 class TestRegistrationIdMessage(TestCase):
